polish changes
