+++ src_c/_pygame.h
@@ -523,7 +523,6 @@ typedef enum {
  * Remember to keep these constants up to date.
  */
 
-//#define PYGAMEAPI_RECT_NUMSLOTS 10
 #define PYGAMEAPI_JOYSTICK_NUMSLOTS 3
 #define PYGAMEAPI_DISPLAY_NUMSLOTS 2
 #define PYGAMEAPI_SURFACE_NUMSLOTS 4
