+++ src_c/include/_pygame.h
@@ -192,6 +192,12 @@ typedef struct pg_bufferinfo_s {
 #define import_pygame_base() IMPORT_PYGAME_MODULE(base)
 #endif /* ~PYGAMEAPI_BASE_INTERNAL */
 
+#ifndef PYGAMEAPI_GEOMETRY_INTERNAL
+
+#define import_pygame_geometry() IMPORT_PYGAME_MODULE(geometry)
+
+#endif /* ~PYGAMEAPI_GEOMETRY_INTERNAL */
+
 typedef struct {
     PyObject_HEAD SDL_Rect r;
     PyObject *weakreflist;
@@ -202,14 +208,16 @@ typedef struct {
     PyObject *weakreflist;
 } pgFRectObject;
 
-
 #define pgRect_AsRect(x) (((pgRectObject *)x)->r)
 #define pgFRect_AsRect(x) (((pgFRectObject *)x)->r)
+
 #ifndef PYGAMEAPI_RECT_INTERNAL
+
 #define pgRect_Type (*(PyTypeObject *)PYGAMEAPI_GET_SLOT(geometry, 0))
 
 #define pgRect_Check(x) ((x)->ob_type == &pgRect_Type)
-#define pgRect_New (*(PyObject * (*)(SDL_Rect *)) PYGAMEAPI_GET_SLOT(geometry, 1))
+#define pgRect_New \
+    (*(PyObject * (*)(SDL_Rect *)) PYGAMEAPI_GET_SLOT(geometry, 1))
 
 #define pgRect_New4 \
     (*(PyObject * (*)(int, int, int, int)) PYGAMEAPI_GET_SLOT(geometry, 2))
@@ -217,7 +225,8 @@ typedef struct {
 #define pgRect_FromObject \
     (*(SDL_Rect * (*)(PyObject *, SDL_Rect *)) PYGAMEAPI_GET_SLOT(geometry, 3))
 
-#define pgRect_Normalize (*(void (*)(SDL_Rect *))PYGAMEAPI_GET_SLOT(geometry, 4))
+#define pgRect_Normalize \
+    (*(void (*)(SDL_Rect *))PYGAMEAPI_GET_SLOT(geometry, 4))
 
 #define pgFRect_Type (*(PyTypeObject *)PYGAMEAPI_GET_SLOT(geometry, 5))
 
@@ -225,13 +234,16 @@ typedef struct {
 #define pgFRect_New \
     (*(PyObject * (*)(SDL_FRect *)) PYGAMEAPI_GET_SLOT(geometry, 6))
 
-#define pgFRect_New4 \
-    (*(PyObject * (*)(float, float, float, float)) PYGAMEAPI_GET_SLOT(geometry, 7))
+#define pgFRect_New4                               \
+    (*(PyObject * (*)(float, float, float, float)) \
+         PYGAMEAPI_GET_SLOT(geometry, 7))
 
-#define pgFRect_FromObject \
-    (*(SDL_FRect * (*)(PyObject *, SDL_FRect *)) PYGAMEAPI_GET_SLOT(geometry, 8))
+#define pgFRect_FromObject                       \
+    (*(SDL_FRect * (*)(PyObject *, SDL_FRect *)) \
+         PYGAMEAPI_GET_SLOT(geometry, 8))
 
-#define pgFRect_Normalize (*(void (*)(SDL_FRect *))PYGAMEAPI_GET_SLOT(geometry, 9))
+#define pgFRect_Normalize \
+    (*(void (*)(SDL_FRect *))PYGAMEAPI_GET_SLOT(geometry, 9))
 
 #define import_pygame_rect() IMPORT_PYGAME_MODULE(geometry)
 #endif /* ~PYGAMEAPI_RECT_INTERNAL */
@@ -501,12 +513,6 @@ typedef struct pgColorObject pgColorObject;
 #define import_pygame_math() IMPORT_PYGAME_MODULE(math)
 #endif /* PYGAMEAPI_MATH_INTERNAL */
 
-#ifndef PYGAMEAPI_GEOMETRY_INTERNAL
-
-#define import_pygame_geometry() IMPORT_PYGAME_MODULE(geometry)
-
-#endif /* ~PYGAMEAPI_GEOMETRY_INTERNAL */
-
 /*
  * Window module
  */
@@ -532,7 +538,6 @@ typedef struct {
  */
 #ifdef PYGAME_H
 PYGAMEAPI_DEFINE_SLOTS(base);
-//PYGAMEAPI_DEFINE_SLOTS(rect);
 PYGAMEAPI_DEFINE_SLOTS(joystick);
 PYGAMEAPI_DEFINE_SLOTS(display);
 PYGAMEAPI_DEFINE_SLOTS(surface);
@@ -546,7 +551,6 @@ PYGAMEAPI_DEFINE_SLOTS(window);
 PYGAMEAPI_DEFINE_SLOTS(geometry);
 #else /* ~PYGAME_H */
 PYGAMEAPI_EXTERN_SLOTS(base);
-//PYGAMEAPI_EXTERN_SLOTS(rect);
 PYGAMEAPI_EXTERN_SLOTS(joystick);
 PYGAMEAPI_EXTERN_SLOTS(display);
 PYGAMEAPI_EXTERN_SLOTS(surface);
