+++ src_c/static.c
@@ -322,7 +322,6 @@ PyInit_pygame_static()
 
 #include "base.c"
 
-#include "rect.c"
 #include "pgcompat_rect.c"
 
 #undef pgSurface_Lock
